.net warnings
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@7201 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ include/MathUtils.h
@@ -153,7 +153,7 @@ protected:
     unsigned int i;
     bzfsrand(0); // ensure consistent seed and same rand value order
     for (i=0; i < 0x10000; i++) {
-      _random_floats[i] = bzfrand();
+      _random_floats[i] = (float)bzfrand();
     }
   }
